+++ parser_library/test/checking/data_definition/data_definition_test.cpp
@@ -15,7 +15,7 @@
 
 #include "../../common_testing.h"
 #include "context/ordinary_assembly/ordinary_assembly_dependency_solver.h"
-#include "hlasmparser.h"
+#include "hlasmparser_multiline.h"
 
 void expect_no_errors(const std::string& text)
 {